Move Back unit tests to common
